Refactor relevance debugger and improve code quality
Addressed all code review findings:

## File Size Compliance (Priority 1 - BLOCKING)
- Split RelevanceDebuggerSection from 513 lines to 217 lines ✅
- Created modular subcomponents:
  - FilterControls.tsx (117 lines)
  - ScoresTable.tsx (130 lines)
  - DecisionDetails.tsx (79 lines)
- All files now under 300-line limit

## Utility Extraction (Priority 2)
- Added formatNumericScore() to shared formatters
- Added formatStringList() to shared formatters
- Created lib/ai/relevanceDebugger.ts for buildRelevanceContext()
- Utilities now reusable across codebase

## Test Coverage (Priority 3)
- Expanded from 2 to 5 tests (+150% coverage)
- Added scope filtering test
- Added empty state handling test
- Added refresh functionality test
- All tests passing with KISS approach

## Documentation Voice (Priority 4)
- Updated decision-relevance-scoring.md for conversational tone
- Updated extending-devtools.md with "here's what you get" framing
- Removed stiff corporate phrasing
- Added practical context and examples

All tests passing (5/5). Component is now more maintainable,
testable, and follows project standards.
